# openapi/systemone.yaml
openapi: 3.1.0
info:
  title: systemone-lite
  version: 0.1.0
  description: >
    Wire-compatible with TypeSafe POST /v1/systemone
    (https://docs.typesafe.ai/api.md). Model quality differs.
servers:
  - url: http://127.0.0.1:8000
paths:
  /v1/systemone:
    post:
      summary: Evaluate state against typed questions
      operationId: systemOne
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemOneRequest"
            example:
              model: systemone-lite-latest
              state: Help! My payouts have been failing for 3 days.
              questions:
                is_urgent:
                  type: noul
                  instructions: Does this convey urgency?
                  criteria:
                    true: Explicitly time-sensitive
                    false: No urgency expressed
                department:
                  type: choice
                  instructions: Which team should handle this?
                  criteria:
                    billing: Payments, invoicing, refunds
                    technical: Bugs, outages, integrations
                    sales: Pricing, upgrades, new accounts
                frustration:
                  type: score
                  instructions: How frustrated is the customer?
                  criteria: [Calm, Frustrated, Very angry]
      responses:
        "200":
          description: Typed answers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemOneResponse"
        "422":
          description: Validation error
  /health:
    get:
      summary: Liveness
      responses:
        "200":
          description: OK
components:
  schemas:
    SystemOneRequest:
      type: object
      additionalProperties: false
      required: [state, questions]
      properties:
        model:
          type: string
          default: systemone-lite-latest
        state:
          oneOf:
            - type: string
            - type: object
            - type: array
        questions:
          type: object
          minProperties: 1
          additionalProperties:
            oneOf:
              - $ref: "#/components/schemas/NoulQuestion"
              - $ref: "#/components/schemas/ChoiceQuestion"
              - $ref: "#/components/schemas/ScoreQuestion"
    NoulQuestion:
      type: object
      additionalProperties: false
      required: [type, instructions]
      properties:
        type:
          const: noul
        instructions:
          type: string
        criteria:
          type: object
          additionalProperties: false
          properties:
            true: { type: string }
            false: { type: string }
    ChoiceQuestion:
      type: object
      additionalProperties: false
      required: [type, instructions, criteria]
      properties:
        type:
          const: choice
        instructions:
          type: string
        criteria:
          type: object
          minProperties: 1
          additionalProperties:
            type: [string, "null"]
    ScoreQuestion:
      type: object
      additionalProperties: false
      required: [type, instructions, criteria]
      properties:
        type:
          const: score
        instructions:
          type: string
        criteria:
          type: array
          minItems: 2
          maxItems: 10
          items:
            type: string
    SystemOneResponse:
      type: object
      additionalProperties: false
      required: [model, answers, usage]
      properties:
        model:
          type: string
        answers:
          type: object
          additionalProperties:
            oneOf:
              - $ref: "#/components/schemas/NoulAnswer"
              - $ref: "#/components/schemas/ChoiceAnswer"
              - $ref: "#/components/schemas/ScoreAnswer"
        usage:
          $ref: "#/components/schemas/Usage"
    NoulAnswer:
      type: object
      additionalProperties: false
      required: [type, noul]
      properties:
        type: { const: noul }
        noul: { type: number, minimum: 0, maximum: 1 }
    ChoiceAnswer:
      type: object
      additionalProperties: false
      required: [type, choice, probabilities, confidence]
      properties:
        type: { const: choice }
        choice: { type: string }
        probabilities:
          type: object
          additionalProperties: { type: number }
        confidence: { type: number, minimum: 0, maximum: 1 }
    ScoreAnswer:
      type: object
      additionalProperties: false
      required: [type, score, legend, probabilities, confidence]
      properties:
        type: { const: score }
        score: { type: number }
        legend:
          type: object
          additionalProperties: { type: string }
        probabilities:
          type: object
          additionalProperties: { type: number }
        confidence: { type: number, minimum: 0, maximum: 1 }
    Usage:
      type: object
      additionalProperties: false
      required: [input_tokens, output_tokens]
      properties:
        input_tokens: { type: integer, minimum: 0 }
        output_tokens: { type: integer, minimum: 0 }